requirements.txt
requests
pdfminer.six
